requirements.txt

Django==5.2.6
djangorestframework==3.15.2
psycopg2-binary
Pillow==10.4.0
django-filter
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
numpy==1.22.4
pytz==2022.1

